reworking the bottom modal UI so it is more reUsable
