+++ app/src/main/java/com/elliottsoftware/calftracker/presentation/sharedViews/BottomNavigations.kt
@@ -31,7 +31,6 @@ data class NavigationItem(
     val weight:Float = 1f,
     val modifier: Modifier = Modifier
 
-
 )
 
 @Composable
